scraper/app/tasks.py
import os
from celery import Celery

# Load environment variables
from dotenv import load_dotenv
load_dotenv()

# Create Celery app
app = Celery('scraper')

# Configure Celery
app.conf.broker_url = os.environ.get('REDIS_URL', 'redis://localhost:6379/0')
app.conf.result_backend = os.environ.get('REDIS_URL', 'redis://localhost:6379/0')

# Configure task schedules
app.conf.beat_schedule = {
    'scrape-blogs-every-hour': {
        'task': 'app.tasks.scrape_blogs',
        'schedule': 3600.0,  # Every hour
    },
}

# Define tasks
@app.task
def scrape_blogs():
    """
    Task to scrape all registered blogs for new content.
    This is a placeholder - implementation will be added in later steps.
    
    Returns:
        dict: Task result summary
    """
    # Placeholder for now
    return {
        "status": "success",
        "blogs_processed": 0,
        "songs_found": 0,
        "errors": []
    } 