scraper/requirements.txt
# Scraper dependencies
beautifulsoup4>=4.12.2
requests>=2.28.2
lxml>=4.9.2

# Task scheduling
celery>=5.2.7
redis>=4.5.4

# Database access
sqlalchemy>=2.0.9
psycopg2-binary>=2.9.6

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.2

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-mock>=3.10.0
factory-boy>=3.2.1
responses>=0.23.1 